Move to typescript.
