+++ package.json
@@ -2,9 +2,10 @@
   "name": "jovani-botquez",
   "version": "1.0.0",
   "description": "Discord bot",
-  "main": "src/index.js",
+  "main": "src/index.ts",
+  "type": "module",
   "scripts": {
-    "dev": "nodemon src/index.js",
+    "dev": "tsx src/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "prettier --write \"**/*.js\""
   },
@@ -29,6 +30,8 @@
     "eslint-plugin-prettier": "^5.1.3",
     "husky": "^9.0.10",
     "prettier": "^3.0.2",
+    "ts-node": "^10.9.2",
+    "tsx": "^4.7.0",
     "typescript": "^5.3.3"
   }
 }
