--- src/commands/resume.js
@@ -1,25 +1,0 @@
-const { GuildMember } = require('discord.js');
-const { useQueue } = require('discord-player');
-const { isInVoiceChannel } = require('../utils/voicechannel');
-
-module.exports = {
-  name: 'resume',
-  description: 'Resume current song!',
-  async execute(interaction) {
-    const inVoiceChannel = isInVoiceChannel(interaction);
-    if (!inVoiceChannel) {
-      return;
-    }
-
-    await interaction.deferReply();
-    const queue = useQueue(interaction.guild.id);
-    if (!queue || !queue.currentTrack)
-      return void interaction.followUp({
-        content: '❌ | No music is being played!'
-      });
-    const success = queue.node.resume();
-    return void interaction.followUp({
-      content: success ? '▶ | Resumed!' : '❌ | Something went wrong!'
-    });
-  }
-};
