--- src/commands/stop.js
@@ -1,23 +1,0 @@
-const { GuildMember } = require('discord.js');
-const { useQueue } = require('discord-player');
-const { isInVoiceChannel } = require('../utils/voicechannel');
-
-module.exports = {
-  name: 'stop',
-  description: 'Stop all songs in the queue!',
-  async execute(interaction) {
-    const inVoiceChannel = isInVoiceChannel(interaction);
-    if (!inVoiceChannel) {
-      return;
-    }
-
-    await interaction.deferReply();
-    const queue = useQueue(interaction.guild.id);
-    if (!queue || !queue.currentTrack)
-      return void interaction.followUp({
-        content: '❌ | No music is being played!'
-      });
-    queue.node.stop();
-    return void interaction.followUp({ content: '🛑 | Stopped the player!' });
-  }
-};
